requirements.txt
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
twilio==9.3.7
python-dotenv==1.0.1
